Support Creating Links in Terminals
This is quite an invasive set of changes, with the aim of also
highlighting keywords in terminal sessions. This adds `terminal` to the
config to support terminals, and raises the minimum vscode version to
1.49.0.

I also took the time to refactor how matching works. We now combine all
the regexps into one, and search for it everywhere, and then when we
have a match, we work out which specific regexp created the match. This
means we only register one link provider per language, and so there is
better control of prioritization between rules. As this logic is fairly
complex, I put it in a new class called MultiRuleMatcher, which both the
LinkProviders subclass. The tests now also use a subclass of
MultiRuleMatcher rather than the DocumentLinkProvider, so they don't
require a partially mocked vscode to run.